Extract DynamoDBCompactionJobStatusStoreTestBase
