+++ java/compaction-status-store/src/test/java/sleeper/compaction/status/job/CompactionStatusStoreTestUtils.java
@@ -21,6 +21,8 @@ import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.StringType;
 
+import java.util.UUID;
+
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.CONFIG_BUCKET;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.FILE_SYSTEM;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.ID;
@@ -38,7 +40,7 @@ public class CompactionStatusStoreTestUtils {
     public static InstanceProperties createInstanceProperties() {
 
         InstanceProperties instanceProperties = new InstanceProperties();
-        instanceProperties.set(ID, "test-instance");
+        instanceProperties.set(ID, UUID.randomUUID().toString());
         instanceProperties.set(CONFIG_BUCKET, "test-bucket");
         instanceProperties.set(FILE_SYSTEM, "test-fs");
         return instanceProperties;
